requirements.txt
flask
authlib
python-dotenv
pymongo
werkzeug
bson
flask-cors
